refactor: reorganize project packages
