+++ app/src/main/java/co/median/android/a2025_theangels_new/data/widget/EmergencyWidget.java
@@ -1,7 +1,7 @@
 // =======================================
 // IMPORTS
 // =======================================
-package co.median.android.a2025_theangels_new.widget;
+package co.median.android.a2025_theangels_new.data.widget;
 
 import android.appwidget.AppWidgetManager;
 import android.appwidget.AppWidgetProvider;
