+++ app/src/main/java/co/median/android/a2025_theangels_new/ui/educations/EducationDetailsActivity.java
@@ -1,7 +1,7 @@
 // =======================================
 // IMPORTS
 // =======================================
-package co.median.android.a2025_theangels_new.educations;
+package co.median.android.a2025_theangels_new.ui.educations;
 
 import android.content.Intent;
 import android.os.Bundle;
