+++ app/src/main/java/co/median/android/a2025_theangels_new/ui/events/active/VolClaimFragment.java
@@ -1,7 +1,7 @@
 // =======================================
 // IMPORTS
 // =======================================
-package co.median.android.a2025_theangels_new.events.active;
+package co.median.android.a2025_theangels_new.ui.events.active;
 
 import android.os.Bundle;
 import android.view.LayoutInflater;
@@ -15,18 +15,18 @@ import androidx.fragment.app.Fragment;
 import co.median.android.a2025_theangels_new.R;
 
 // =======================================
-// VolStatusFragment - Displays volunteer's event progress/status
+// VolClaimFragment - Fragment for the volunteer claim stage
 // =======================================
-public class VolStatusFragment extends Fragment {
+public class VolClaimFragment extends Fragment {
 
     // =======================================
-    // onCreateView - Inflates the layout for volunteer status UI
+    // onCreateView - Inflates the layout for volunteer claim UI
     // =======================================
     @Nullable
     @Override
     public View onCreateView(@NonNull LayoutInflater inflater,
                              @Nullable ViewGroup container,
                              @Nullable Bundle savedInstanceState) {
-        return inflater.inflate(R.layout.fragment_vol_status, container, false);
+        return inflater.inflate(R.layout.fragment_vol_claim, container, false);
     }
 }
