+++ app/src/main/java/co/median/android/a2025_theangels_new/ui/events/active/VolStatusFragment.java
@@ -1,7 +1,7 @@
 // =======================================
 // IMPORTS
 // =======================================
-package co.median.android.a2025_theangels_new.events.active;
+package co.median.android.a2025_theangels_new.ui.events.active;
 
 import android.os.Bundle;
 import android.view.LayoutInflater;
@@ -15,18 +15,18 @@ import androidx.fragment.app.Fragment;
 import co.median.android.a2025_theangels_new.R;
 
 // =======================================
-// VolClaimFragment - Fragment for the volunteer claim stage
+// VolStatusFragment - Displays volunteer's event progress/status
 // =======================================
-public class VolClaimFragment extends Fragment {
+public class VolStatusFragment extends Fragment {
 
     // =======================================
-    // onCreateView - Inflates the layout for volunteer claim UI
+    // onCreateView - Inflates the layout for volunteer status UI
     // =======================================
     @Nullable
     @Override
     public View onCreateView(@NonNull LayoutInflater inflater,
                              @Nullable ViewGroup container,
                              @Nullable Bundle savedInstanceState) {
-        return inflater.inflate(R.layout.fragment_vol_claim, container, false);
+        return inflater.inflate(R.layout.fragment_vol_status, container, false);
     }
 }
