+++ app/src/main/java/co/median/android/a2025_theangels_new/ui/home/HomeActivity.java
@@ -32,10 +32,10 @@ import co.median.android.a2025_theangels_new.data.services.EventStatusDataManage
 import co.median.android.a2025_theangels_new.data.models.EventType;
 
 import co.median.android.a2025_theangels_new.R;
-import co.median.android.a2025_theangels_new.map.MapFragment;
+import co.median.android.a2025_theangels_new.data.map.MapFragment;
 import co.median.android.a2025_theangels_new.data.models.UserSession;
 import co.median.android.a2025_theangels_new.data.models.Event;
-import co.median.android.a2025_theangels_new.events.list.RecentEventsAdapter;
+import co.median.android.a2025_theangels_new.ui.events.list.RecentEventsAdapter;
 
 // =======================================
 // HomeActivity - Displays the home screen and handles location permission logic
